# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Meowgram API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: posts
    description: Post operations
  - name: ratings
    description: Rating operations
  - name: likes
    description: Like operations
  - name: comments
    description: Comment operations
  - name: users
    description: User operations
  - name: Storage
    description: Object storage upload and serving endpoints.

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"

  /login:
    get:
      operationId: beginBrowserLogin
      tags: [auth]
      summary: Start browser OIDC login
      responses:
        "302":
          description: Redirect to OIDC

  /callback:
    get:
      operationId: handleBrowserLoginCallback
      tags: [auth]
      summary: Complete browser OIDC login
      responses:
        "302":
          description: Redirect after login

  /logout:
    get:
      operationId: logoutBrowserSession
      tags: [auth]
      summary: Clear session and OIDC logout
      responses:
        "302":
          description: Redirect to OIDC logout

  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange mobile OIDC code for session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"

  /mobile-auth/logout:
    post:
      operationId: logoutMobileSession
      tags: [auth]
      summary: Delete mobile session
      responses:
        "200":
          description: Logout processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutMobileSessionResponse"

  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get all posts (feed)
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [new, top, trending]
          description: Sort order
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized

  /posts/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [posts]
      summary: Get top rated posts leaderboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsResponse"

  /posts/{postId}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /posts/{postId}/rate:
    post:
      operationId: ratePost
      tags: [ratings]
      summary: Rate a post (1-10)
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RatePostBody"
      responses:
        "200":
          description: Rating submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized

  /posts/{postId}/like:
    post:
      operationId: toggleLike
      tags: [likes]
      summary: Toggle like on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResponse"
        "401":
          description: Unauthorized

  /posts/{postId}/comments:
    get:
      operationId: getComments
      tags: [comments]
      summary: Get comments for a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentsResponse"
    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUploadUrlBody'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestUploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
        '500':
          description: Failed to serve object.

  /users/{userId}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile with posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]

    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"

    GetCurrentAuthUserResponse:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"

    ExchangeMobileAuthorizationCodeBody:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string

    LogoutMobileSessionResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        username:
          type: string
        userAvatar:
          type: string
          nullable: true
        imageUrl:
          type: string
        caption:
          type: string
          nullable: true
        mediaType:
          type: string
          enum: [image, video]
        averageRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
        likeCount:
          type: integer
        commentCount:
          type: integer
        userRating:
          type: integer
          nullable: true
        userLiked:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - imageUrl
        - mediaType
        - ratingCount
        - likeCount
        - commentCount
        - userLiked
        - createdAt

    PostsResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
      required:
        - posts
        - total

    RatePostBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - score

    RatingResponse:
      type: object
      properties:
        averageRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
        userRating:
          type: integer
      required:
        - ratingCount
        - userRating

    LikeResponse:
      type: object
      properties:
        liked:
          type: boolean
        likeCount:
          type: integer
      required:
        - liked
        - likeCount

    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: string
        username:
          type: string
        userAvatar:
          type: string
          nullable: true
        text:
          type: string
        createdAt:
          type: string
      required:
        - id
        - postId
        - userId
        - username
        - text
        - createdAt

    CommentsResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required:
        - comments

    AddCommentBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text

    UserProfileResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            profileImage:
              type: string
              nullable: true
            createdAt:
              type: string
          required:
            - id
            - username
            - createdAt
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        totalPosts:
          type: integer
      required:
        - user
        - posts
        - totalPosts

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    CreatePostBody:
      type: object
      required: [imageUrl, mediaType]
      properties:
        imageUrl:
          type: string
          description: Object path from GCS upload (e.g. /objects/uploads/uuid)
        caption:
          type: string
          nullable: true
        mediaType:
          type: string
          enum: [image, video]

    RequestUploadUrlBody:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1

    RequestUploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/RequestUploadUrlBody'